Continuação

+++ Project/src/GeneralRepository/Bet.java
@@ -2,36 +2,31 @@ package GeneralRepository;
 
 public class Bet {
 
-    public int SpectatorID;         // SpectatorID identifica um espectador
-    public int horseID;             // HorseID identifica um cavalo
-    public double Betvalue;         // Betvalue identifica um valor de uma aposta
+    public int SpectatorID;       // SpectatorID identifica um espectador
+    public int horseID;           // HorseID identifica um cavalo
+    public int Betvalue;         // Betvalue identifica um valor de uma aposta
 
-    
     public int getSpectatorID() {
         return SpectatorID;
     }
 
-    
     public void setSpectatorID(int SpectatorID) {
         this.SpectatorID = SpectatorID;
     }
-
     
     public int getHorseID() {
         return horseID;
     }
-
     
     public void setHorseID(int horseID) {
         this.horseID = horseID;
     }
-
     
-    public double getBetvalue() {
+    public int getBetvalue() {
         return Betvalue;
     }
 
-    public void setBetvalue(double Betvalue) {
+    public void setBetvalue(int Betvalue) {
         this.Betvalue = Betvalue;
     }
 }
